+++ src/com/company/courseOne/homework3/Task2.java
@@ -23,7 +23,7 @@ package com.company.courseOne.homework3;
  */
 public class Task2
 {
-    private static final float IOS = 0;
+    private static final int IOS = 0;
     private static final int ANDROID = 1;
 
     public static void main(String[] args)
